Move arrow transformers into a separate module
